refactor: split wamr & sys modules into separate files for clarity
